[GOHW-2]: Covert to kopecks

+++ internal/model/currency_rate.go
@@ -2,5 +2,5 @@ package model
 
 type CurrencyRate struct {
 	Name  string
-	Value float64
+	Value int64
 }
